flashall: use pre-shipped fastboot when available
If we don't set $FASTBOOT, flashall.sh always uses the distro provided
"fastboot" executable.
This is problematic, because older distros (such as Ubuntu 18.04) have a
very old version of "fastboot".

To avoid compatibility issues on flashing, use the prebuilt fastboot
binary if it's available.

The prebuilt binary is shipped with our CI builds.

Note: android 12's fastboot (31.0) is not capable of formatting in ext4,
so an erase will be done for partitions such as metadata.

Test: flashall.sh shows locally used fastboot and boot to UI.
Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>