device: take wakelock on zygote restart
When the framework is taken down (with adb shell stop), the system
immediately tries to enter suspend due to auto-suspend.

However, suspend is un-supported on our platforms and the android
framework should prevent us from suspending due to the
def_stay_on_while_plugged_in overlay.

Since autosuspend kicks in so fast, the android system cannot block
suspend when zygote is restarting.

Take a wakelock while zygote restarts to fix this.

This is based on an AOSP solution from aosp/master [1] and should
improve overall VTS stability according to [2].

Note: that we can't unlock the wakelock on zygote=running because that's
too early in the boot flow.

[1] https://android-review.googlesource.com/c/platform/system/core/+/2353993
[2] https://github.com/GloDroid/glodroid_device/commit/1cde03e969dcd12fb0caedb8a41205e0a871b30b
      adb shell start # home screen comes back up

Test: adb shell stop # board does not enter suspend
Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>