device: add android.hardware.hardware_keystore feature
Add android.hardware.hardware_keystore feature and set its value to 3

This feature is used to check Keymaster/Keymint version and it should
be equal to the version reported by Keymaster or Keymint

The optee based keymaster 3.0 returns a value of 3. Thus, set the value
to 3.

This fixes the following cts test:
Module: CtsKeystoreTestCases
Test: android.keystore.cts.KeyAttestationTest#testAttestationKmVersionMatchesFeatureVersion

Signed-off-by: Safae Ouajih <souajih@baylibre.com>