all: fix developer settings crash by enabling bluetooth
When opening developer option in the Settings app, a crash is observed
because we don't have the bluetooth feature enabled.

Even though recent AM62x SK EVM versions do not chip with a Bluetooth
chip, we still want to enable it in Android to avoid this crash.

Enable it via the AOSP provided btlinux HAL.

Note: this generates some crashes in the BT HAL while booting but
eventually stops

Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>
[mkorpershoek: reworded commit msg]
Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>