am62x_lp: remove am62x_lp lunch target
With android 13 it's no longer a problem to not have a wifi interface
even if the wifi hal is included.

The am62x_lp lunch target was mainly created for this purpose, so we
can drop it.

Remove am62x_lp lunch target create just to disable wifi and pick

different bootloaders.

Update flash.sh script to support multiple board with '--board' arg.

Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>