BoardConfig: add display modules to BOARD_VENDOR_RAMDISK_KERNEL_MODULES
In recovery, we only ship the modules listed in
BOARD_VENDOR_RAMDISK_KERNEL_MODULES.

Since that list does not contain the display related modules, we don't
have any display support in recovery:

[    8.100207][   T96] recovery: Failed to read max brightness: No such file or directory
[    8.108450][   T96] recovery: Screensaver disabled
[   13.869491][   T96] recovery: timeout wait for 'graphics' took 5000ms
[   14.120908][   T99] recovery (99) used greatest stack depth: 13096 bytes left
[   14.121710][   T96] recovery: Starting recovery (pid 96) on Thu Jan  1 00:00:15 1970
[   14.139026][   T96] recovery: locale is [en-US]

Fix it by adding the missing modules.

* tidss, sii902x and display-connector are needed for HDMI.
* panel-simple is needed for LVDS panel.

Note: since tidss is now part of the ramdisk, it is probed before
pvrsrvkm, so it becomes /dev/dri/card0.

Test: adb reboot recovery # we can see recoveryUI
      # boot to regular launcher
Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>