camera: move location front -> back to fix capture
When taking a picture with the com.android.camera2 app, the resulting picture
seems wrongly encoded. In fact, even if the size is correct, the result seems
very pixelated (like it's a thumbnail picture).

This is because we miss the following fix in the camera app [1].

Since we don't want to fork our camera package, switch to a back camera instead.
The back camera does not have this problem.

[1] https://android-review.googlesource.com/c/platform/packages/apps/Camera2/*/2263681
Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>