am62p: ueventd: make /dev/video* accessible to media and camera
Commit aa0d7f3e94eb ("am62p: ueventd: split out /dev/video* for media/camera")
seperated /dev/video[0,1] from the other video devices under the assumption that
these first two are the wave5 driver.

This is not necessarily true, as wave5.ko is probed *after* UVC when a UVC cam
is plugged before booting.

Fix this by ensure that *all* /dev/video* nodes can be access by both the media
user (for v4l2_code2c) and the camera group (for the camera HALs).

Test: dumpsys media.camera shows a webcam
Fixes: aa0d7f3e94eb ("am62p: ueventd: split out /dev/video* for media/camera")
Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>