am62x: set ro.config.per_app_memcg to false
With limited memory, OTA is broken when we apply it via update_engine.
Indeed when we switch to the new slot, snapuserd_proxy fails and the
device reboot:

  libprocessgroup: Failed to get stats for /dev/memcg/apps: No such file or directory
  libprocessgroup: Failed to make and chown /dev/memcg/apps/uid_0: No such file or directory
  init: Could not start snapuserd_proxy: createProcessGroup(0, 170, 0) failed for service 'snapuserd_proxy': No such file or directory
  init: Service 'snapuserd_proxy' failed to start due to a fatal error

When ro.config.low_ram=true and ro.config.per_app_memcg property is
not set, UsePerAppMemcg will return true and so snapuserd_proxy will
try to access to: /dev/memcg/apps*

However these required directories are not yet present.

A patch has been made to fix this issue:
https://android-review.googlesource.com/c/platform/system/core/+/2820943

This fix is not yet present in our AOSP.
As workaround we manually set: ro.config.per_app_memcg=false
Thus UsePerAppMemcg will return false and snapuserd_proxy won't try to
access to /dev/memcg/apps*.

This is a workaround, once the proper fix is present in our AOSP, we
should revert this change.

Signed-off-by: Julien Masson <jmasson@baylibre.com>