am62p|am62x: activate logpersist.cat by default
In userdebug, set the correct property and add the corresponding
rule in selinux in order to activate logpersist.cat by default.
Thanks to that a persitent logcat will always be available from
the start of the system even when a crash occurs.
This allows the boottime scripts to gather all the timings even
when a crash occurs.

Signed-off-by: Nicolas Belin <nbelin@baylibre.com>