common: Remove thermal from framework_compatibility_matrix
The thermal hal we use is HIDL version 2.0.

This is no longer compatible with Android:

ERROR: files are incompatible: android.hardware.thermal@2.0::IThermal/default is deprecated in compatibility matrix at FCM Version 8; it should not be served.
    because it matches android.hardware.thermal@2.0::IThermal/default from /system/etc/vintf/compatibility_matrix.5.xml
android.hardware.thermal@2.0::IThermal/default is deprecated in compatibility matrix at FCM Version 8; it should not be served.
    because it matches android.hardware.thermal@2.0::IThermal/default from /system/etc/vintf/compatibility_matrix.7.xml
android.hardware.thermal@2.0::IThermal/default is deprecated in compatibility matrix at FCM Version 8; it should not be served.
    because it matches android.hardware.thermal@2.0::IThermal/default from /system/etc/vintf/compatibility_matrix.6.xml
The following instances are in the device manifest but not specified in framework compatibility matrix:
    android.hardware.thermal@1.0::IThermal/default
    android.hardware.thermal@2.0::IThermal/default
Suggested fix:
1. Update deprecated HALs to the latest version.
2. Check for any typos in device manifest or framework compatibility matrices with FCM version >= 8.
3. For new platform HALs, add them to any framework compatibility matrix with FCM version >= 8 where applicable.
4. For device-specific HALs, add to DEVICE_FRAMEWORK_COMPATIBILITY_MATRIX_FILE or DEVICE_PRODUCT_COMPATIBILITY_MATRIX_FILE.: Success
INCOMPATIBLE

Remove it.

In the next commit, we will use the newly written AIDL HAL instead.

Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>