shared: audio: policy: Rename default output to Speaker
The AIDL audio HAL does not accept "Wired Headphones" in the
<attachedDevices> section:

> E AHAL_Config: convertDevicePortsInModuleToAidl: Review Audio Policy config:
>   <attachedDevices> list is incorrect or devicePort "Wired Headphones" type=
>   AudioDeviceDescription{type: OUT_HEADPHONE, connection: analog} is incorrect.

> F AHAL_Config: Function: convertModuleConfigToAidl Line: 397 Failed result (BAD_VALUE)

In fact, it explicitely dis-allows *all* ports having a "connection"
in convertDevicePortsInModuleToAidl().

Since OUT_HEADPHONE has an analog connection, we can't use it in the
<attachedDevices> section.

Removing "Wired Headphones" would solve the policy crash but would
require us to implement jack detection (which we don't have).

Switch the default output type to Speaker instead, which solves the issue.

Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>