common: Update Audio HAL to AIDL
Replace TinyHAL with the newly developed AIDL Audio HAL which is
based on the reference implementation from:
//hardware/interfaces/audio/aidl/default

The new audio HAL includes:
- primary
- remote submix
- usb
- bluetooth
- effects

We no longer need to list the hals in the device manifest because it
comes via a fragment.

Note: Bluetooth audio is mandatory, otherwise the board won't boot.

Test: playback on jack # works, same volume as before
Test: atest VtsHalAudioCoreTargetTest
Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>