common: declare gadgethal service in init.am62x.usb.rc
The init-gadgethal-sh service is no longer declared
in the usb hal after migration to apex.
The update to the android revision 'android-15.0.0_r7'
caused errors due to the script being located
outside of the apex.

Declare the service in the device init.am62x.usb.rc file
to solve the issue.

Signed-off-by: Maxime Fournier <mfournier@baylibre.com>