common: Enable activities_on_secondary_displays for all targets
Historically, we only added this to the car variants to allow users
to choose between mirroring and extended display.

With the Android 15 launcher, display is *always* in mirroring, even
when this permission is included.

In order to launch activities on secondary displays, we can use alternative
launchers such as MD Launcher (development/samples/MultiDisplay).

For more information on Multi Display, see [1].

[1] https://source.android.com/docs/core/display/multi_display/testing-dev-environment
Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>