am67a: config: fix dfu configuration file name
The wrong file was being copied from config folder to the out folder
with the right name. This caused the dfu configuration to being wrong
and either flashing the wrong bootloader or failing to find the right
one.

Copy the am67a config file instead of the am62px one.

Signed-off-by: Maxime Fournier <mfournier@baylibre.com>